refactor: Phase 1 Type Dehydration - Move validation logic to internal/validation (#47)
- Extract Validate(), validateEnumeratedDim(), validateHierarchicalDim()
- Extract IsReservedColumnName() and IsValidPrefix()
- Update imports in nanostore/types.go to use validation package
- Maintain all existing functionality with passing tests
